refactor: deployment configuration types and add in Spark
