+++ packages/deploy-configurations/types/deployment-config/aave-like-protocol.ts
@@ -1,0 +1,1 @@
+export type AaveLikeProtocol = 'PoolDataProvider' | 'LendingPool' | 'Oracle'
