+++ packages/deploy-configurations/types/deployment-config/ajna-protocol.ts
@@ -1,0 +1,37 @@
+import { ConfigEntry } from './config-entries'
+
+export type AjnaProtocol =
+  | 'AjnaPoolInfo'
+  | 'AjnaProxyActions'
+  | 'AjnaPoolPairs_ETHDAI'
+  | 'AjnaPoolPairs_ETHUSDC'
+  | 'AjnaPoolPairs_RETHDAI'
+  | 'AjnaPoolPairs_RETHETH'
+  | 'AjnaPoolPairs_RETHUSDC'
+  | 'AjnaPoolPairs_USDCETH'
+  | 'AjnaPoolPairs_USDCWBTC'
+  | 'AjnaPoolPairs_USDCDAI'
+  | 'AjnaPoolPairs_WBTCDAI'
+  | 'AjnaPoolPairs_WBTCUSDC'
+  | 'AjnaPoolPairs_WSTETHDAI'
+  | 'AjnaPoolPairs_WSTETHETH'
+  | 'AjnaPoolPairs_WSTETHUSDC'
+  | 'AjnaPoolPairs_CBETHETH'
+  | 'AjnaPoolPairs_TBTCWBTC'
+  | 'AjnaPoolPairs_TBTCUSDC'
+  | 'AjnaPoolPairs_ETHGHO'
+  | 'AjnaPoolPairs_WSTETHGHO'
+  | 'AjnaPoolPairs_GHODAI'
+  | 'AjnaPoolPairs_RETHGHO'
+  | 'AjnaPoolPairs_WBTCGHO'
+  | 'AjnaPoolPairs_CBETHGHO'
+  | 'AjnaPoolPairs_WLDUSDC'
+  | 'AjnaPoolPairs_USDCWLD'
+  | 'AjnaPoolPairs_SDAIUSDC'
+  | 'AjnaPoolPairs_YFIDAI'
+  | 'AjnaPoolPairs_YIELDETHETH'
+  | 'AjnaPoolPairs_YIELDBTCWBTC'
+  | 'AjnaRewardsManager'
+  | 'AjnaRewardsClaimer'
+  | 'ERC20PoolFactory'
+export type AjnaProtocolContracts = Record<AjnaProtocol, ConfigEntry>
