+++ packages/deploy-configurations/types/deployment-config/automation.ts
@@ -1,0 +1,4 @@
+import { ConfigEntry } from './config-entries'
+
+export type Automation = 'AutomationBot' | 'AutomationBotV2' | 'AutomationBotAggregator'
+export type AutomationContracts = Record<Automation, ConfigEntry>
