+++ packages/deploy-configurations/types/deployment-config/common.ts
@@ -1,0 +1,35 @@
+import { ConfigEntry } from './config-entries'
+import { Tokens } from './tokens'
+
+export type Common =
+  | Tokens
+  | 'UniswapRouterV3'
+  | 'BalancerVault'
+  | 'OneInchAggregator'
+  | 'AuthorizedCaller'
+  | 'FeeRecipient'
+  | 'GnosisSafe'
+  | 'MerkleRedeemer'
+  | 'DssCharter'
+  | 'DssProxyActions'
+  | 'DssProxyActionsCharter'
+  | 'DssMultiplyProxyActions'
+  | 'DssCropper'
+  | 'DssProxyActionsCropjoin'
+  | 'DssProxyActionsDsr'
+  | 'Otc'
+  | 'OtcSupportMethods'
+  | 'ServiceRegistry'
+  | 'GuniProxyActions'
+  | 'GuniResolver'
+  | 'GuniRouter'
+  | 'CdpRegistry'
+  | 'DefaultExchange'
+  | 'NoFeesExchange'
+  | 'LowerFeesExchange'
+  | 'LidoCrvLiquidityFarmingReward'
+  | 'ChainlinkPriceOracle_USDCUSD'
+  | 'ChainlinkPriceOracle_ETHUSD'
+  | 'SdaiOracle'
+
+export type CommonContracts = Record<Common, ConfigEntry>
