+++ packages/deploy-configurations/types/deployment-config/index.ts
@@ -1,0 +1,2 @@
+export { ConfigEntry, Contracts, SystemContracts } from './config-entries'
+export { SystemConfig } from './deployment-config'
