+++ packages/deploy-configurations/types/deployment-config/maker-protocol.ts
@@ -1,0 +1,120 @@
+import { ConfigEntry } from './config-entries'
+
+export type MakerProtocol =
+  | 'FlashMintModule'
+  | 'Chainlog'
+  | 'CdpManager'
+  | 'GetCdps'
+  | 'Jug'
+  | 'Pot'
+  | 'End'
+  | 'Spot'
+  | 'Dog'
+  | 'Vat'
+  | 'McdGov'
+
+export type MakerProtocolJoins =
+  | 'MCD_JOIN_DAI'
+  | 'MCD_JOIN_ETH_A'
+  | 'MCD_JOIN_ETH_B'
+  | 'MCD_JOIN_ETH_C'
+  | 'MCD_JOIN_BAT_A'
+  | 'MCD_JOIN_USDC_A'
+  | 'MCD_JOIN_USDC_B'
+  | 'MCD_JOIN_PSM_USDC_A'
+  | 'MCD_JOIN_TUSD_A'
+  | 'MCD_JOIN_WBTC_A'
+  | 'MCD_JOIN_WBTC_B'
+  | 'MCD_JOIN_WBTC_C'
+  | 'MCD_JOIN_ZRX_A'
+  | 'MCD_JOIN_KNC_A'
+  | 'MCD_JOIN_MANA_A'
+  | 'MCD_JOIN_USDT_A'
+  | 'MCD_JOIN_PAXUSD_A'
+  | 'MCD_JOIN_PSM_PAX_A'
+  | 'MCD_JOIN_COMP_A'
+  | 'MCD_JOIN_LRC_A'
+  | 'MCD_JOIN_LINK_A'
+  | 'MCD_JOIN_BAL_A'
+  | 'MCD_JOIN_YFI_A'
+  | 'MCD_JOIN_GUSD_A'
+  | 'MCD_JOIN_PSM_GUSD_A'
+  | 'MCD_JOIN_UNI_A'
+  | 'MCD_JOIN_RENBTC_A'
+  | 'MCD_JOIN_AAVE_A'
+  | 'MCD_JOIN_MATIC_A'
+  | 'MCD_JOIN_WSTETH_A'
+  | 'MCD_JOIN_WSTETH_B'
+  | 'MCD_JOIN_DIRECT_AAVEV2_DAI'
+  | 'MCD_JOIN_GUNIV3DAIUSDC1_A'
+  | 'MCD_JOIN_GUNIV3DAIUSDC2_A'
+  | 'MCD_JOIN_CRVV1ETHSTETH_A'
+  | 'MCD_JOIN_UNIV2DAIETH_A'
+  | 'MCD_JOIN_UNIV2WBTCETH_A'
+  | 'MCD_JOIN_UNIV2USDCETH_A'
+  | 'MCD_JOIN_UNIV2DAIUSDC_A'
+  | 'MCD_JOIN_UNIV2ETHUSDT_A'
+  | 'MCD_JOIN_UNIV2LINKETH_A'
+  | 'MCD_JOIN_UNIV2UNIETH_A'
+  | 'MCD_JOIN_UNIV2WBTCDAI_A'
+  | 'MCD_JOIN_UNIV2AAVEETH_A'
+  | 'MCD_JOIN_UNIV2DAIUSDT_A'
+  | 'MCD_JOIN_RWA001_A'
+  | 'MCD_JOIN_RWA002_A'
+  | 'MCD_JOIN_RWA003_A'
+  | 'MCD_JOIN_RWA004_A'
+  | 'MCD_JOIN_RWA005_A'
+  | 'MCD_JOIN_RWA006_A'
+  | 'MCD_JOIN_RETH_A'
+  | 'MCD_JOIN_GNO_A'
+
+export type MakerProtocolPips =
+  | 'PIP_ETH'
+  | 'PIP_BAT'
+  | 'PIP_USDC'
+  | 'PIP_WBTC'
+  | 'PIP_TUSD'
+  | 'PIP_WETH'
+  | 'PIP_ZRX'
+  | 'PIP_KNC'
+  | 'PIP_MANA'
+  | 'PIP_USDT'
+  | 'PIP_PAXUSD'
+  | 'PIP_PAX'
+  | 'PIP_COMP'
+  | 'PIP_LRC'
+  | 'PIP_LINK'
+  | 'PIP_BAL'
+  | 'PIP_YFI'
+  | 'PIP_GUSD'
+  | 'PIP_UNI'
+  | 'PIP_RENBTC'
+  | 'PIP_AAVE'
+  | 'PIP_MATIC'
+  | 'PIP_WSTETH'
+  | 'PIP_ADAI'
+  | 'PIP_UNIV2DAIETH'
+  | 'PIP_UNIV2WBTCETH'
+  | 'PIP_UNIV2USDCETH'
+  | 'PIP_UNIV2DAIUSDC'
+  | 'PIP_UNIV2ETHUSDT'
+  | 'PIP_UNIV2LINKETH'
+  | 'PIP_UNIV2UNIETH'
+  | 'PIP_UNIV2WBTCDAI'
+  | 'PIP_UNIV2AAVEETH'
+  | 'PIP_UNIV2DAIUSDT'
+  | 'PIP_GUNIV3DAIUSDC1'
+  | 'PIP_GUNIV3DAIUSDC2'
+  | 'PIP_CRVV1ETHSTETH'
+  | 'PIP_RWA001'
+  | 'PIP_RWA002'
+  | 'PIP_RWA003'
+  | 'PIP_RWA004'
+  | 'PIP_RWA005'
+  | 'PIP_RWA006'
+  | 'PIP_RETH'
+  | 'PIP_GNO'
+
+export type MakerProtocolCommonContracts = Record<MakerProtocol, ConfigEntry>
+export type MakerProtocolJoinContracts = Record<MakerProtocolJoins, ConfigEntry>
+export type MakerProtocolPipContracts = Record<MakerProtocolPips, ConfigEntry>
