+++ packages/deploy-configurations/types/deployment-config/spark-protocol.ts
@@ -1,0 +1,5 @@
+import { AaveLikeProtocol } from './aave-like-protocol'
+import { ConfigEntry } from './config-entries'
+
+export type SparkProtocol = AaveLikeProtocol
+export type OptionalSparkProtocolContracts = Partial<Record<SparkProtocol, ConfigEntry>>
