+++ packages/deploy-configurations/types/deployment-config/system-keys.ts
@@ -1,0 +1,9 @@
+export enum SystemKeys {
+  MPA = 'mpa',
+  COMMON = 'common',
+  AAVE = 'aave',
+  SPARK = 'spark',
+  MAKER = 'maker',
+  AUTOMATION = 'automation',
+  AJNA = 'ajna',
+}
