+++ packages/deploy-configurations/types/deployment-config/tokens.ts
@@ -1,0 +1,56 @@
+export type Tokens =
+  | 'AAVE'
+  | 'ADAI'
+  | 'BAL'
+  | 'BAT'
+  | 'CBETH'
+  | 'COMP'
+  | 'CRVV1ETHSTETH'
+  | 'DAI'
+  | 'ETH'
+  | 'GHO'
+  | 'GNO'
+  | 'GUNIV3DAIUSDC1'
+  | 'GUNIV3DAIUSDC2'
+  | 'GUSD'
+  | 'KNC'
+  | 'LDO'
+  | 'LINK'
+  | 'LRC'
+  | 'MANA'
+  | 'MATIC'
+  | 'PAX'
+  | 'PAXUSD'
+  | 'RENBTC'
+  | 'RETH'
+  | 'RWA001'
+  | 'RWA002'
+  | 'RWA003'
+  | 'RWA004'
+  | 'RWA005'
+  | 'RWA006'
+  | 'SDAI'
+  | 'STETH'
+  | 'TBTC'
+  | 'TUSD'
+  | 'UNI'
+  | 'UNIV2AAVEETH'
+  | 'UNIV2DAIETH'
+  | 'UNIV2DAIUSDC'
+  | 'UNIV2DAIUSDT'
+  | 'UNIV2ETHUSDT'
+  | 'UNIV2LINKETH'
+  | 'UNIV2UNIETH'
+  | 'UNIV2USDCETH'
+  | 'UNIV2WBTCDAI'
+  | 'UNIV2WBTCETH'
+  | 'USDC'
+  | 'USDT'
+  | 'WBTC'
+  | 'WETH'
+  | 'WLD'
+  | 'WSTETH'
+  | 'YIELDBTC'
+  | 'YIELDETH'
+  | 'YFI'
+  | 'ZRX'
