+++ packages/dma-library/src/utils/abis/get-abi-for-contract.ts
@@ -11,16 +11,15 @@ import aaveV3PriceOracleOptimismABI from '@abis/external/protocols/aave/v3-l2/aa
 import aaveV3ProtocolDataProviderOptimismABI from '@abis/external/protocols/aave/v3-l2/aaveProtocolDataProvider.json'
 import aaveV3PoolOptimismABI from '@abis/external/protocols/aave/v3-l2/pool.json'
 // Other imports
+import { AaveLikeProtocol } from '@deploy-configurations/types/deployment-config'
 import { Network } from '@deploy-configurations/types/network'
 import { getForkedNetwork as coalesceNetwork } from '@deploy-configurations/utils/network/index'
 import { Protocol } from '@dma-library/types'
 import { ContractInterface } from '@ethersproject/contracts'
 import { providers } from 'ethers'
 
-export type AllowedContractNames = 'poolDataProvider' | 'pool' | 'oracle'
-
 export async function getAbiForContract(
-  contractName: AllowedContractNames,
+  contractName: AaveLikeProtocol,
   provider: providers.Provider,
   protocol: Protocol,
 ): Promise<ContractInterface> {
@@ -42,35 +41,35 @@ export async function getAbiForContract(
 }
 
 const abiByContractName: Partial<
-  Record<Network, Partial<Record<Protocol, Record<AllowedContractNames, ContractInterface>>>>
+  Record<Network, Partial<Record<Protocol, Record<AaveLikeProtocol, ContractInterface>>>>
 > = {
   [Network.MAINNET]: {
     AAVE: {
-      poolDataProvider: aaveV2PoolDataProviderABI,
-      pool: aaveV2PoolABI,
-      oracle: aaveV2OracleABI,
+      PoolDataProvider: aaveV2PoolDataProviderABI,
+      LendingPool: aaveV2PoolABI,
+      Oracle: aaveV2OracleABI,
     },
     AAVE_V3: {
-      poolDataProvider: aaveV3ProtocolDataProviderABI,
-      pool: aaveV3PoolABI,
-      oracle: aaveV3PriceOracleABI,
+      LendingPool: aaveV3PoolABI,
+      PoolDataProvider: aaveV3ProtocolDataProviderABI,
+      Oracle: aaveV3PriceOracleABI,
     },
     Spark: {
-      poolDataProvider: aaveV3ProtocolDataProviderOptimismABI,
-      pool: aaveV3PoolOptimismABI,
-      oracle: aaveV3PriceOracleOptimismABI,
+      PoolDataProvider: aaveV3ProtocolDataProviderOptimismABI,
+      LendingPool: aaveV3PoolOptimismABI,
+      Oracle: aaveV3PriceOracleOptimismABI,
     },
   },
   [Network.OPTIMISM]: {
     AAVE_V3: {
-      poolDataProvider: aaveV3ProtocolDataProviderOptimismABI,
-      pool: aaveV3PoolOptimismABI,
-      oracle: aaveV3PriceOracleOptimismABI,
+      PoolDataProvider: aaveV3ProtocolDataProviderOptimismABI,
+      LendingPool: aaveV3PoolOptimismABI,
+      Oracle: aaveV3PriceOracleOptimismABI,
     },
     Spark: {
-      poolDataProvider: aaveV3ProtocolDataProviderOptimismABI,
-      pool: aaveV3PoolOptimismABI,
-      oracle: aaveV3PriceOracleOptimismABI,
+      PoolDataProvider: aaveV3ProtocolDataProviderOptimismABI,
+      LendingPool: aaveV3PoolOptimismABI,
+      Oracle: aaveV3PriceOracleOptimismABI,
     },
   },
 }
